requirements.txt
pandas>=2.2.2
numpy>=1.26.4
tushare>=1.2.89
python-dotenv>=1.0.1
tqdm>=4.66.5
